requirements.txt
aiohttp==3.7.4.post0
asgiref==3.4.1
async-timeout==3.0.1
attrs==21.2.0
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
charset-normalizer==2.0.6
discord-py-interactions==3.0.2
discord.py==1.7.3
Django==3.2.7
djangorestframework==3.12.4
flake8==3.9.2
gunicorn==20.1.0
humanize==3.11.0
idna==3.2
importlib-metadata==4.8.1
mccabe==0.6.1
multidict==5.1.0
psycopg2==2.9.1
pycodestyle==2.7.0
pycparser==2.20
pyflakes==2.3.1
PyNaCl==1.4.0
pytz==2021.1
requests==2.26.0
six==1.16.0
sqlparse==0.4.2
typing-extensions==3.10.0.2
urllib3==1.26.7
yarl==1.6.3
zipp==3.6.0
